formatStorageDomain removes the VG too.
When removing a SD the VG should be removed too (atomicity).
Actually this is what RHEV-M does. Therefore the functions
split is redundant.
The removeVG function is marked as DEPRECATED (to be removed).
The VG is removed in this version in order to minimize the
chance to remain with a VG thats not belongs to a SD.

Change-Id: Ie993abd2a812f1ecec127cc637794994591e48ea
Reviewed-on: http://gerrit.usersys.redhat.com/659
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>