Related to BZ#713173 - Add killProc rollback to deleteImage
When we run deleteImage with post-zeroing flag, we need to issue 'dd' for this.
The rollback procedure for deleteImage it's actually roll forward.
In this situation we need first to kill previous 'dd' operation that may hold
file handler (or LV) open and prevent us to run the new deleteImage.
Note that in case we need to run this roll forward action on different host (SPM switchover)
the 'dd' killing will do nothing and still may fail.

Change-Id: I1407626ab06583f5507f3aaa8ebf2a4e4edd60f9
Reviewed-on: http://gerrit.usersys.redhat.com/707
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>