BZ#726630 Use only os.access to check permissions
The old implementation of fileUtils.pathExists was using a flawed
re-implementation of os.access as backup check for files with a NFS
stale handle, the consequence was that accessible files were
reported as non-accessible.
We now use os.stat to refresh the NFS handle and then os.access to
determine weather we can read and write the file.

Change-Id: Ic7b892886416e866178ac88c04cbfb68ed0c055b
Reviewed-on: http://gerrit.usersys.redhat.com/733
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Jakub Libosvar <libosvar@redhat.com>