BZ#727871: guestIF: rewrite the message reading code.
The previous code handle only one message at a time. This mean that in case of
multiple messages vdsm handled only the first message while the following
messages was not handle until more data was send from the guest.

This patch change this behavior to scan the entire buffer for messages and
handle all of them.

Change-Id: Id3c2ef7f08cadd686329a280c06f7db438f667db
Reviewed-on: http://gerrit.usersys.redhat.com/785
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>