BZ#736422: A new JSON-based protocol with the guest's agent.
In order to recover from a malformed message send from the guest's
agent, a new JSON-based protocol was implemented. This allow vdsm
to drop bad messages (messages are terminated with a '\n' character)
and handle the following good messages.

This patch also include code related to BZ#729254 as the new
protocol includes the disk usage information.

Change-Id: Ie6096b326cc58879c8f63b72e32586657fffeca0
Reviewed-on: http://gerrit.usersys.redhat.com/807
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>