Make the README file more useful
It is not important to have a full GPLv2+ boilerplate in the
README file. It can just refer to the main COPYING file and
note the OpenSSL exception.

More important for the README is info about what VDSM is, how
to install it, package it and how to report bugs / get help,
etc

* README: Update content to be more useful

Change-Id: I24bd2db470be0ef59f5f6af501da52f5d25d5fb7
Reviewed-on: http://gerrit.usersys.redhat.com/841
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Tested-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>