BZ#736034 - Add metadataignore switch to pvcreate.
We would like to pre-allocate metadata space in all PVs
used for our VGs. However, we are using only the metadata
of the first PV.
This is important for an occasion where the first PV has
to be removed and the metadata it carries has to be copied
into somewhere else.

Could not have been done without Haim's help.

Change-Id: Ib040902610fd2c5b0de04de95b57f0bfbdd1e11e
Reviewed-on: http://gerrit.ovirt.org/711
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>