BZ#725967 - Have global pool users use the limiter as well
In the last patch in the quest for using only 1 storage pool I left out
the global pool users. Because the global pool was initialized on import
because after this patch all the handlers are created at once this
caused all the handlers to be created on import. This made the pool act
strange on certain occasions. The fact that the global users didn't pass
through the limiter was a mistake in itself.

This patch:
* Moves ProcessPoolLimiter to the processPool module because it is
  the correct place and I didn't want all the global pool users
  importing sd.py.
* Removes ProcessPoolDict and creates a new class called
  ProcessPoolMultiplexer that does the same thing but does not inherit
  from dict which means you don't get the __set__ methods and other
  garbage. It also moves the logic to the processPool module.
* Adds getProcessPool() to outOfProcess so there is convenient access
  to the global multiplexer.
* Adds getGlobalProcPool() to outOfProcess so global pool users can get
  the global pool on demand and not on import
* Changes all calls to the global process pool to go through the
  aforementioned methods so that the pool is never created on import

Change-Id: Id52f872b3bfd12088a601f69678a602d4c6f3c76
Reviewed-on: http://gerrit.usersys.redhat.com/910
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>