BZ#733669 Improve error reporting for wrong local hostname
This patch improves the error reporting when migration fails.
It is mainly designed to return the new error 'wrongHost' to
the manager when the migration destination has a wrong local
hostname (which will cause the migration to fail without any
informative message).
Additionally it checks the local hostname during the bootstrap
process and produces a warning if needed.

In this patch:
* add local hostname check during bootstrap (report a warning)
* return the new error 'wrongHost' if migration destination has
  a wrong local hostname
* source node must return a specific error when migration fails

Change-Id: Iaf0ae2d62772b01d05d9d5e62d2433c1758d69b1
Reviewed-on: http://gerrit.usersys.redhat.com/924
Tested-by: David Naori <dnaori@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>