BZ#736316 persist network config on rhevh, too.
On RHEV-H we should not really worry about backing up old configuration
files, since they are already stored under
/config/etc/sysconfig/network-scripts. However, vdsm-store-net-config
looks under the netconfbackup dir for the files it needs to store - when it
sees nothing there, nothing is persisted for next boot.

Quickest fix is to make configNetwork._persistentBackup() hold a copy of
ifcfg-* in netconfbackup for rhevh, too.

Change-Id: I69f6e218cce1d9075920e4b16014ea6d15db292c
Reviewed-on: http://gerrit.usersys.redhat.com/926
Tested-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>