BZ#748222 netinfo: report operstate, not adminstate
Until commit 4153e08acb8ae23ef79 (fixing BZ#726359), when reporting if a
network inteface is up or down, Vdsm was considering the operative status.
That was the correct behavior, since users do not really care that they have
configered the nic to be up while the line is cut off by evil mice (or quality
engineers).

Change-Id: I32b74a57d96ba0feea4816749c6c45ea34695d2a
Reviewed-on: http://gerrit.usersys.redhat.com/1057
Reviewed-by: Igor Lvovsky <ilvovsky@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>