BZ#754445 - Separate granting from callback emitting
Granting has to happen inside the request transaction to prevent races
with cancel\timeouts. The callback is code that the subsystem does not
know anything about. For that reason it is crucial to run the callback
outside of the lock scope. To keep both demands I split up emit() from
grant().

Change-Id: Ic12d170e64399e37a555960c03804778ad7d053b
Reviewed-on: http://gerrit.ovirt.org/254
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>