Avoid to activate parent volume on cloning
The benefit is that we can now create new snapshot volumes without
looking at the parent. This is indispensable for live snapshots since
the parent volume is in use by the VM.

WARNING: if you run an SPM host with this change you need all the
HSM hosts to have: I4b5cb55a868cb9bf5a720610273af139ee858a1a (Don't
lock the image when preparing the VM path) and you won't be able use
the domain from old SPMs/HSMs anymore.

In this patch:
* Do not activate the parent volume on cloning (it could be in use
  by the VM), the only information needed is the size and it can be
  provided as additional parameter in qemu-img
* Do not change the volume permission to read-only when setting the
  parent to "internal" (it could be in use by the VM)

Change-Id: I8a7d488da5ab0c27fc2e083976d4bc44e6f810e3
Reviewed-on: http://gerrit.ovirt.org/570
Reviewed-by: Igor Lvovsky <ilvovsky@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>