Do not lock the image when preparing the VM path
Since the resource manager scope is local there is no need for a VM
to take an exclusive lock on the images: in fact the SPM running on
an other machine could still manipulate such images.
Using the resource manager in general on HSM is wrong because it is
mapped on actions that are valid only on the SPM such as: modifing
the lvm permissions and the modification time in the metadata.

In this patch:
* add an activate/deactivate image method to StorageDomain
* implement the activate/deactivate image for BlockStorageDomain
* add the prepareImage/teardownImage methods in HSM

Change-Id: I4b5cb55a868cb9bf5a720610273af139ee858a1a
Reviewed-on: http://gerrit.ovirt.org/554
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>