Add an object model to clientIF
Create an object-based API for vdsm.  The current objects are: Task,
StoragePool, StorageDomain, Image, Volume, ISCSIConnection, LVMVolumeGroup, and
VM.  Additionally, a Global object currently handles APIs which don't map nicely
to one of the above objects.  Right now this interface is only an abstraction
that maps directly to the flat function namespace.  This will change later in
the patch series and more over time.  Since this API is internal (it only
affects bindings), we can continue to change it over time without affecting
ovirt-engine.

Change-Id: I2ee54277c87a02ac8dafe29a8761e2acf0f7397f
Signed-off-by: Adam Litke <agl@us.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/984
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>