Move API out of clientIF
Many API functions are implemented inside clientIF but this is no longer the
proper place for them.  Move API-related functions into their proper objects
inside API.py.  clientIF still owns the storage dispatcher and the vm list.  In
the future, I would like to move these out of clientIF as well but I have
decided to reserve that for a future patch series.

Change-Id: I94ad036cfaf3d1bec2f5b4049979a1bf0e2810ce
Signed-off-by: Adam Litke <agl@us.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/985
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>