getDevInfo: report partitioned devices, too
One cannot create a PV on a partitioned device, and therefor such
devices where not reported to Engine. This proved surprising to users
who wonder where their LUN disappeared.

Vdsm should report all devices, and ovirt-engine should mark partitioned
devices as unworthy of a PV. In the future, Vdsm may allow to forcefully
remove a partition table from a device, to make it usable as a PV.

Change-Id: Ia5eddfcb587aec5337adaef810bf11665a7dffde
Reviewed-on: http://gerrit.ovirt.org/944
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>