Add before_vm_destroy hook
Run a hook directory when a Vm is asked to die. Note that a Vm may die
without being asked to, were the hook would not run.

This hook is currently missing from Vdsm since I was not sure what
should be its exact semantic. However in hindsight, we'd better have
something (even awkward) and improve after feedback from the real world.

Change-Id: Ide2c143921e58944df97c3dc74cd20cd85b1dae8
Reviewed-on: http://gerrit.ovirt.org/1175
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>