BZ#773371 call `vdsmd reconfigure` after bootstrap
If installation change ssl from False to True and added valid
certificates, we need to reconfigure vdsm to use it.

This was done by the "reconfigure" verb for our SysV daemon, but
unsupported by systemd.

This patch is an ugly hack for an oVirt release, until we have a proper
vdsm-tool for these purposes.

Change-Id: Ide0b9c1fb86c8d7bcea54d1ba8c8d9a7598af733
Reviewed-on: http://gerrit.ovirt.org/1360
Tested-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>