API: Separate clientIF and the xmlrpc server
Currently the xmlrpc server functionality is entwined with the clientIF class.
In order to proceed with the API bridge plan (http://ovirt.org/wiki/Vdsm_API) I
would like to separate the xmlrpc server into its own class.  This brings
clientIF one step closer to becoming a generic API Bridge that is able to host
additional bindings such as REST and QMF.

Despite the size of this patch, this was a pretty simple refactoring.  I made
the following choices as I split this out:

- The xmlrpc server binding class has start() and prepareForShutdown() APIs
- clientIF will still own the storage dispatcher but API wrapper mechanisms
  have moved into the xmlrpc class.

Change-Id: I5026e11f3d39343c957b59606e90c23acbb598ba
Signed-off-by: Adam Litke <agl@us.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/620
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>
Tested-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>