task.py: avoid overtly silly casting to str()
For a reason unknown (to me), task.py does not use pickle to persist its
data to file. We cannot fix that without breaking
backward-compatibility. However, we can avoid pointless casts to str,
and can encode data to utf8 before storing to disk.

Note that this patch does not solve BZ#567077 as the storage subsystem
prevents entry of Unicode arguments. This prevention must be kept
for old storage domain formats.

Change-Id: I97db724db436e99b95c3524e70fab540f7191334
Reviewed-on: http://gerrit.ovirt.org/1622
Tested-by: Haim Ateya <hateya@redhat.com>
Reviewed-by: Eduardo <ewarszaw@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>