Refactor storage connection subsystem
Currently connection management is an abysmal mess. This is a
transitional patch
retrofitting the old API to some of the concepts we intend to implement.

Notice:
Permissions are no longer checked on storage connect only on domain
initialization. This is good as the storage connection is far from an
appropriate place for that check.

I've removed stale root folder handling for NFS. I will add it back I
can't do everything at once.

There are also no proper engine acceptable exceptions yet

Change-Id: Ifc5c160aba9b4cc88225dde0c44f9b766c6a4c0d
Reviewed-on: http://gerrit.ovirt.org/1038
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>