Adding flowID
This patch will implement the flowID schema in vdsm side. The idea is
provide a mechanism to identify the flow between ovirt engine and vdsm API calls
in order to identify the entire scope of the backend action
(which possibly consisted of several VSDM calls). At moment, we are
going to use HTTP headers to implement this feature to keep the
compability with current API, avoiding possible breakes.

Change-Id: I6c818d7a90fea41663fd48620ac9718451a8662f
Reviewed-on: http://gerrit.ovirt.org/1221
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>