vm payload: add file injection to vm
this patch adds new parameter to create verb vmPayload,
user can send file content and have them in the vm in floppy (.vmf) or
cdrom (.iso) format.

vmPayload is in specParams
'vmPayload': {'file': {'filename': 'content' }}
for now we are only supporting file type in cdrom or floppy,
we intend to add more formats such as network url which vdsm will
download the file and place it in cdrom or floppy.

Change-Id: I256475342c79690a95ad999335522f99714cdc8b
Signed-off-by: Shahar Havivi <shaharh@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/2321
Tested-by: Shahar Havivi <shavivi@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>