BZ#790014 Fix multiple issues in detachSD
In this patch:
* Remove __unmountLastMaster, the manager cannot deactivate the master
  storage domain (where the SPM resource is hold)
* Move all the checks related to the master migration inside masterMigrate
* detachSD cannot be used to migrate a master

Note that this patch makes it possible to deactivate a master domain
(thus breaking the pool) even if there are other active domains in the
pool. A future patch should discuss if reverting to the old semantics is
necessary, and if so - how it should be done.

Change-Id: I04a7c16df86b828c514daab34f21e352919217f9
Reviewed-on: http://gerrit.ovirt.org/1876
Tested-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>