boostrap completion: fix no reboot function and report.
This patch fixes the functionality of no reboot as a
given argument to the script. So using 0 will avoid
rebooting and anything else will continue rebooting
as it did so far.

Additionaly the vds_bootsrap_complete XML report was corrected;
when the script is being asked not to reboot the machine,
so the operation is still considered to be a success.

Change-Id: I89c441fb999c696f60e1ed3172c0ed57649e78bb
Reviewed-on: http://gerrit.ovirt.org/2121
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Roy Golan <rgolan@redhat.com>
Tested-by: Roy Golan <rgolan@redhat.com>