boostrap completion: Restart vdsmd service.
Recently we've fixed the functionality which avoids
reboot once bootstrap is completed. However, the vdsmd
service requiers restart in order to pick up the new
configuration and the new certificates issues during
the bootstrap.

This patch uses the same logic as the ovirt-node reboot
function, which reboots the machine or restarts the
vdsmd servide when reboot is avoided.

Change-Id: Ib95b7cd9aeeff89c0b56e10c6b34f5b853d1f848
Reviewed-on: http://gerrit.ovirt.org/2194
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Doron Fediuck <dfediuck@redhat.com>