Use certtool to generate the certificates
Using certtool is easier (commands are easy to understand and maintain),
requires less steps (at least 1 less: the crt request), and it's the
tool suggested by libvirt in the documentation (this should limit the
mismatches). Achieving the same thing with openssl would have required
shipping a openssl.cnf and having a temporary certificate request to be
self-signed.

Change-Id: I73fcfa2389a276bc1110921a509ef99a408318dd
Reviewed-on: http://gerrit.ovirt.org/2657
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Tested-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Igor Lvovsky <ilvovsky@redhat.com>