VDSM interface for the virt-alignment-scan tool from libguestfs
The libguestfs-tools rpm has an application called virt-alignment-scan.
http://libguestfs.org/virt-alignment-scan.1.html
Virt-alignment-scan scans the partition alignment in guest VMs and reports
if a partition in guest VM is not aligned.

This is a vdsm interface to virt-alignment-scan. It is now possible
to scan the alignment of VM guest partitions directly from vdsm.
In the future, it should be possible to expose this functionality
through vdsClient.

Change-Id: I2d2c146a0fa3101317720f2e9c373e0d21b5cfdf
Signed-off-by: Saša Tomić <tomic80@gmail.com>
Reviewed-on: http://gerrit.ovirt.org/2916
Reviewed-by: Igor Lvovsky <ilvovsky@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>