BZ#842631 Use domain proxies instead of actual domain references
Because of issues where we had 2 instances of the same domain causing
trouble we changed the invalidation scheme to only invalidate when no
one is holding a reference to the domain object. This causes a lot of
correctnes errors, espescially on not manager hosts. Having a proxy in
the middle will allow us to invalidate the domain object and have all
reference owners access the correct object.

Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Change-Id: Ifdedd601ca25e6cac3f1f68f2917776401c0d838
Reviewed-on: http://gerrit.ovirt.org/3044
Tested-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Ayal Baron <abaron@redhat.com>