Orthogonal storage repository conversion
Currently domain format conversion is considered an "upgrade". This is
not correct as in the future we may allow conversion from one domain
type to another without a definite supercedence value.

For instance, one domain format might be better suited for quick
image manipulation while another will be better suited for space saving
image manipulations.

Also in the future we will introduce the new image repositories that
will replace the current domain scheme. As these are completely
different objects there needs to bee a 3rd party to convert to\from
them.

This also separates the conversion process from the domain logic
allowing simpler more concise handling of the conversion process.

Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Change-Id: I7f43faaa9578ddafbae8e3aa01c02b1d42b177ab
Reviewed-on: http://gerrit.ovirt.org/3045
Reviewed-by: Ayal Baron <abaron@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Tested-by: Federico Simoncelli <fsimonce@redhat.com>