Define some constants in API.py
The API bindings should be implemented in terms of the API bridge (API.py) and
not rely on vdsm internals.  Currently the xmlrpc binding imports some internal
modules in order to provide some default values (ie. BLANK_UUID and volume
constants).  Constants that are needed at the API level need to appear in
API.py.

This patch converts the obvious offenses.  If this approach is acceptible I will
follow up with additional patches.  In the future, APIs that take 'magic'
parameters must formally define the valid values in API.py.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Change-Id: Iaef19f956ec1d6f83fd48e71a2f0a018505266be
Reviewed-on: http://gerrit.ovirt.org/2262
Reviewed-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-by: Peter V. Saveliev <peet@redhat.com>
Tested-by: Peter V. Saveliev <peet@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>