Add pep8 checks to build process
This adds pep8 verification check. pep8 is the recommended styling for
python. The pep8 tester doesn't check all of the style recomendations
but it does help. Currently most of our code doesn't conform, so in
order to integrate it slowly there is a whitelist declared in
/Makefile.am.
When fixing a file to conform please add the file to the whitelist so
that after the patch is submitted the file will be tracked for
conformance.

More about pep8
http://www.python.org/dev/peps/pep-0008/

Change-Id: I77a2ae06467a676cf8b9574474852071599117ac
Reviewed-on: http://gerrit.ovirt.org/3259
Reviewed-by: Peter V. Saveliev <peet@redhat.com>
Tested-by: Peter V. Saveliev <peet@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>