BindingXMLRPC: Check irs init status
During vdsm initialization, it is possible that IRS will fail to initialize and
will instead be None.  During API refactoring we added a TODO to the code to
re-enable checking for this case.  In this patch, we check if cif.irs is None
and if so, activate a default storage error message for all storage methods.

Change-Id: I87b1509f4cb95319d0a88cec75f941d61aa5dae5
Signed-off-by: Adam Litke <agl@us.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/3571
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>