Separate the Volume.share implementation
The current Volume.share method is exposing the hardlink/symlink option
which is an implementation detail of the domain type (hardlink for file
domains and symlinks for block domains). This makes any external (or
generic) use of such call domain-type dependant.
This patch separates part of the Volume.share implementation to use the
correct link type transparently.

In this patch:
* Separate the Volume.share implementation for block and file domains
* Separate the shareVolumeRollback implementation
* Link also the lease file on file domains (when required)

Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Change-Id: I965b56dc98fe2d6db3ef6f3d7efe2a7b9791b555
Reviewed-on: http://gerrit.ovirt.org/3737
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Ayal Baron <abaron@redhat.com>