Introduce a new verb setBalloonTarget
We need this API to achieve memory overcommitment on guest level. Here's
the definition of the new API:

setBalloonTarget:
    Dynamically change the target amount of physical memory allocated
    to a guest.

    Input parameter:
        vmId   - VM UUID
        target - target physical memory size in kibibytes
    Return value:
        success: return doneCode
        failure: return errCode including underlying libvirt error message.

And the ballooning information is added into the response of
getVmStats(). It can be collected on vdsm client by this way:

   balloon_info = s.getVmStats(vmId)['statsList'][0]['balloonInfo']

Please note that setBalloonTarget can work only when a virtio balloon
device is available for the guest. So to test it with vdsm API, you need
put the following python dictionary into the 'devices' list in VM's conf
to enable balloon device:

{'device': 'memballoon', 'type': 'balloon', 'specParams':
 {'model': 'virtio'}}

On ovirt-engine, there's checkbox 'Memory Balloon Enabled' in the
'New/Edit VM' UI. You can enable it in cluster 3.1.

Change-Id: I36b1211dbce28bb5bb5926f2db7e2e42a698efc9
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/3669
Reviewed-by: Adam Litke <agl@us.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>