Implement Popen without forking back to python
Python's implementation of Popen forks back to python before execing.
Forking a python proc is a very complex and volatile process and we
already have hacks to try and avoid some of the bugs involved with it.

This is a simpler method of execing that doesn't go back to python after
forking. This allows for faster safer exec.

Should solve bugs involving with GIL\GC\Logging deadlocks related to the
python implementation.

This also makes means that python-devel and gcc are needed for testing
as the extension needs to be compiled to be tested.

Change-Id: Ief4ae42eb57551d7f5893bf562da451a9c98f219
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/3944
Reviewed-by: Ayal Baron <abaron@redhat.com>
Reviewed-by: Wenchao Xia <xiawenc@linux.vnet.ibm.com>
Tested-by: Wenchao Xia <xiawenc@linux.vnet.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>