Avoid skipping libvirt when add bridgeless network
Change since V2
    - Adjust subject and comment to make it more reasonable.

Change since V1
    - Move this validation into _addNetworkValidation based
      on Dan's suggestion.

The bridgeless network can not be added when the options contain
skip libvirt. When I dig into the code, seems this type of network
is supported by libvirt, but it will not do its real job if skip
libvirt based on current logical in addNetwork. So set limitation
to avoid such situation.

Change-Id: Ia61fab631e8eecf3dc698a94d72858d7a72ade08
Signed-off-by: Lei Li <lilei@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/3919
Reviewed-by: Shahar Havivi <shavivi@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>