Let LC_ALL=C when running tests
The function validateDDBytes() in vdsm/storage/misc.py needs C locale.
When vdsm runs /bin/dd, the stderr output of /bin/dd is inspected by
validateDDBytes(). Then the last line of the output is splitted by space
and the 0th word is the total count of bytes transferred. In a locale
like zh_CN.utf8(Chinese), the stderr output message is different from
that in C locale. Although dd runs successfully, validateDDBytes() can
not take out the count of the bytes transferred, and an exception
will be raised.

When vdsmd is started by vdsmd.init, LC_ALL is set to C to provide a
compatible locale for vdsmd. However, when running tests, the locale is
not set to C. So "make check" will fail in a zh_CN.utf8 locale in all
tests related to dd, and this causes "autobuild.sh" and "make rpm" fail
as well. So, set LC_ALL=C will provide a standard locale for running the
tests.

Change-Id: I93176e8948aea11cb6adf11d8dc295e43bb1dfee
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/4243
Reviewed-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>