storage: Move nfsSD.getFileUtils() to FileStorageDomain class
If I create an iso-domain out of a LOCALFS storage type, when I
call getIsoList on the storage pool which includes the iso-domain
vdsm complains that LocalFS does not have getFileList method.

The nfsSD class has the exact method for finding files that is
needed.  By moving the implementation to the base class for both
nfs and localfs types, they can share the implementation.

Change-Id: I31cf133464751ffcf222140e83fa8d0e32940c05
Signed-off-by: Ryan Harper <ryanh@us.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/4478
Reviewed-by: Ewoud Kohl van Wijngaarden <ewoud@kohlvanwijngaarden.nl>
Reviewed-by: Adam Litke <agl@us.ibm.com>
Reviewed-by: Eduardo <ewarszaw@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>