vdsm: using model and index for controller address
This patch uses the index and model of controllers when saving the alias
and address fields. In some cases, like the USB controllers, there is
more than one controller, and the address is important as different USB
controllers should have different addresses. So, using the
device+index+model is a unique ID for the USB controllers.

Change-Id: I68dbdedd3b3d5b1dacf232d2f5b5bb3c59f6d565
Signed-off-by: Oved Ourfali <oourfali@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/4499
Reviewed-by: Igor Lvovsky <ilvovsky@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>