Change to use WatchedFileHandler instead of FileHandler in logger.conf
We find that no new log message appears in log file after manually changing
the vdsm log when vdsm is running. We have to restart vdsm to get logging
work again. It could happen when people try to add some markers to pinpoint
the related messages during toubleshooting. It's caused by the log file gets
associated with a new inode after the manual change. To make logging survive
the change, use WatchedFileHandler instead of FileHandler.

In addition, WatchedFileHandler was provided specially for the external
log rotate tools. It can notice that the log file change performed by
logrotate, so we don't need to have the option 'copytruncate' in
vdsm-logrotate.conf any more.

Change-Id: Iaf1cd7f3c7d568b8351eaa3ce262532caec94c7e
Reported-by: Changming Bai <baichm@linux.vnet.ibm.com>
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/4463
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Changming Bai <baichm@linux.vnet.ibm.com>