tests: Clean up sample text in fileUtilTests.py
I understand the value of humor and I am certain no harm was intended but I
must take issue with some of the sample text being used in the fileUtilTests
module.  I know the quotes were intended as jokes on their original
programs, but I find them offensive, subjugative to women, and totally
inappropriate for source code.  I know our project wishes to be welcoming to
all developers and therefore I request that this text be removed in favor of
something neutral such as Lorem ipsum.  Thank you for your understanding.

Change-Id: Ie6339620f2d0692a390fb082a3e4512d31d78f2c
Signed-off-by: Adam Litke <agl@us.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/4517
Reviewed-by: Ryan Harper <ryanh@us.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>