tests: Package up tests
This patch packages up the test suite so that it can be installed on a target
machine to test the installed vdsm.  The check-local target still runs the tests
against the files in the source directory.  When installed, the tests will use
the system installed version.  This paves the way for adding some functional
tests that will run against the running vdsm instance.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Change-Id: I40a5d97b8cffcbaa465cab61aa890ff2520a4179
Reviewed-on: http://gerrit.ovirt.org/4451
Reviewed-by: Ewoud Kohl van Wijngaarden <ewoud@kohlvanwijngaarden.nl>
Reviewed-by: Wenchao Xia <xiawenc@linux.vnet.ibm.com>
Tested-by: Wenchao Xia <xiawenc@linux.vnet.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>