BZ#820525 lvm: ignore future additions to attr bits
lvm2-2.02.95 has added few more bits to LV_ATTR_BITS, and that broke
vdsm. In this patch we try to be more future-compatible, and ignore bits
that are added on top of the ones we know about.

This patch assumes that lvm2 would never delete existing bits, reorder
them, or add new ones at the beginning. It is sent per the request of
Alasdair Kergon of lvm2 fame
https://bugzilla.redhat.com/show_bug.cgi?id=820525#c4

Change-Id: Ie3cd1ad3e211076c0fa2b3554af0baf72d8ac45c
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/4565
Reviewed-by: Ayal Baron <abaron@redhat.com>
Tested-by: Peter V. Saveliev <peet@redhat.com>
Reviewed-by: Peter V. Saveliev <peet@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>