rest-api: Allow optional data in DELETE request bodies
By default, cherrypy does not allow data in DELETE HTTP requests.  In our
API, we would like to pass parameters to some DELETE requests.  We
accomplish this by configuring two things:

1) Set request.methods_with_bodies to include DELETE.  This tells cherrypy
that we should process the body of DELETE requests too.

2) Create a hook to enable body processing only for DELETE requests with an
appropriate decorator.  This makes a request body optional for the
indicated functions.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Change-Id: I65466894358029c7fe5f0ef46a32492682975761
Reviewed-on: http://gerrit.ovirt.org/5049
Reviewed-by: Dan Kenigsberg <danken@redhat.com>