rest-api: Add unit tests
Add a framework and unit tests for the REST API.  These tests override the vdsm
API bridge so that the REST code can be tested in isolation.  A functional test
suite that verifies the operation of both the xmlrpc and rest bindings against a
running vdsmd is still needed.

This test module uses some tricky magic to dynamically-generate a fake API.py
that will return pre-programmed return values to the REST API.  Using this
technique, a cherrypy server can be tested with an unmodified Controller.

The following files are created/modified:
 restData.py   - Contains expected data for:
                 Faked API calls, REST requests, REST responses
 restTests.py  - Contains the actual test cases
 testrunner.py - Modified to import some additional needed modules

Signed-off-by: Adam Litke <agl@us.ibm.com>
Change-Id: Ife07553718315136bbb5a74558dd189dee4e475f
Reviewed-on: http://gerrit.ovirt.org/3757
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>