Change storageServer to handle numeric connection values
engine is sending the connection information as a map of
string keys and string values. VDSM tries to parse version,retrans and
timeout as numbers, but due to the nature of the passed map
it receives a string containing numerical values for these arguments, and
fails.
This patch performs cast to int of the numercial values, and in case of
failure, throws an exception indicating which value raised the exception

Change-Id: I4d1716a6f66a108d60f39fad14d92625a692d0e3
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/5062
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Ayal Baron <abaron@redhat.com>