BZ#808874 Skip master validation on repoStat
The current master validation code is affected by two different race
conditions triggered by both startSpm and stopSpm.

The first and clearer race condition is triggered by the master status
(fetched asynchronously by the domainMonitor thread) that could be
several seconds old. Comparing such status with the current SPM role
(SPM/HSM) during a transition (startSpm, stopSpm) is wrong.

The second race condition (more difficult to expose) happens again
during the startSpm/stopSpm methods since the mount/unmount of the
master domain is not atomic with the switch of the SPM role.

There are two correct ways to solve this issue:

1. Maintain the SPM role in the domainMonitor thread so that it can be
atomically paired with the mount status. To do so we would need to
introduce two new methods to signal to the domainMonitor the beginning
and the end of a transition (startSpm, stopSpm) and its final outcome
(role = SPM/HSM).

2. Keep track of the latest SPM/HSM transition (eg: sp.lastRoleChange)
and verify that the statistics are recent enough before validating the
master mount (eg: st.lastCheck > sp.lastRoleChange).

The code involved in this issue was introduced in 8d400ac (BZ#518840)
to add an automatic recovery system for those pools that were missing
the master mount/directory (preventing the SPM to start).
Eventually that error could be reported during startSpm but it's not
related to repoStats (which is anychronous to startSpm and therefore
race prone).

At the moment the engine handles the StorageDomainMasterError error code
issuing a reconstructMaster command.

Implementing one of the two fixes proposed above just to maintain this
validation is an overdesign given that the two problematic situations
(SMP-master-not-valid and HSM-master-mounted) should be addressed in a
different way by VDSM and not reported to the engine.

Consequences for the engine:
* The error code StorageDomainMasterError will never be received
* The engine won't issue the automatic reconstructMaster command (in
  this particular case)

Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Change-Id: I837877b5431edc3cfc6648e8fb654df2291413fc
Reviewed-on: http://gerrit.ovirt.org/5134
Reviewed-by: Ayal Baron <abaron@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Haim Ateya <hateya@redhat.com>